Completed Ex4_1.js
Created inner Style Object for Obj Literal as it turns out .attribute.width is not for divs.

Good news, found that you can use the window object to call methods, with the caveat it only works on global objects.

Modified code to reflect shifting style properties to a style inner object and the calling of style methods with global Window Object in array format rather than typical method format.